requirements.txt
flask
psycopg2-binary
matplotlib
python-dotenv
schedule
numpy
pandas